requirements.txt
streamlit
torch
transformers
gtts
Pillow
scikit-learn
matplotlib